requirements.txt
pyyaml
numpy
jupyter
gym
catkin_pkg
rospkg
h5py
keras
pandas
tensorflow
torchvision
matplotlib
